+++ src/data/layout/linked/_reexport_alloc.rs
@@ -1,0 +1,8 @@
+// devela/src/data/layout/linked/_reexport_alloc.rs
+
+use crate::{_reexport, _tags};
+
+_reexport! { rust: alloc::collections, location: "data/layout/linked", tag: _tags!(data_structure),
+    doc: "A doubly-linked list with owned nodes.",
+    LinkedList
+}