+++ src/data/layout/ord/generic.rs
@@ -1,4 +1,4 @@
-// devela/src/data/layout/sort/generic.rs
+// devela/src/data/layout/ord/generic.rs
 //
 //! Implements sorting algorithms for exclusive generic arrays `[T: Ord; N]`.
 //