+++ src/data/layout/ord/impls_alloc.rs
@@ -1,4 +1,4 @@
-// devela/src/data/layout/sort/impls_alloc.rs
+// devela/src/data/layout/ord/impls_alloc.rs
 //
 //! Implements sorting algorithms for exclusive generic arrays `[T: Ord; N]`.
 //