+++ src/data/layout/ord/primitives.rs
@@ -1,4 +1,4 @@
-// devela/src/data/layout/sort/primitives.rs
+// devela/src/data/layout/ord/primitives.rs
 //
 //! Implements compile-time sorting algorithms for arrays of primitives.
 //