+++ src/data/topol/graph/mod.rs
@@ -1,0 +1,28 @@
+// devela/src/data/topol/graph/mod.rs
+//
+#![doc = crate::_DOC_DATA_TOPOL_GRAPH!()] // public
+#![doc = crate::_doc!(modules: crate::data::topol; graph)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: collections)]
+//!
+//
+
+// #[cfg(test)]
+// mod _test;
+// #[cfg(any(test, feature = "_docs_examples"))]
+// mod _example;
+
+// mod adj;
+// mod csr;
+// mod impls;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     adj::_all::*,
+        //     // csr::_all::*,
+        // };
+        // #[cfg(any(test, feature = "_docs_examples"))]
+        // pub use super::_example::*;
+    }
+}