+++ src/data/topol/linked/_reexport_alloc.rs
@@ -1,8 +1,0 @@
-// devela/src/data/topol/linked/_reexport_alloc.rs
-
-use crate::{_reexport, _tags};
-
-_reexport! { rust: alloc::collections, location: "data/topol", tag: _tags!(data_structure),
-    doc: "A doubly-linked list with owned nodes.",
-    LinkedList
-}